refactor(commands): extract /export command into export.rs
